Review and update Debian copyright file
Add entries for osx/queue.h, that was introduced in version
2.7.0, and linux/tree.h, that was introduced all the way back
in version 1.0.0 but somehow never got included in the Debian
copyright file.

Tweak license texts to make them applicable to more cases,
using http://spdx.org/licenses/ as the source - it basically
boils down to use "the copyright holder" instead of eg. "the
OpenBSD Foundation". Reflow them a bit while at it.

Review the whole file to make sure the all information is
complete and accurate.