Update Debian copyright file
Upstream believes that spectrwm no longer contains a significant
enough part of dwm's code to warrant including the relative
license text and copyright notices in the source.

This means we can avoid having a separate stanza for spectrwm.c
in the Debian copyright file, and can simply use the package-wide
license and copyright notices.