Upgrade x-terminal-emulator from Recommends: to Depends:
The original rationale for turning the terminal emulator into a
weak dependency was that it had to be done for the screen locking
program, so might as well piggyback on that effort.

Now that the default screen locking program has been switched to
slock, part of suckless-tools, there is no point in not checking
the default programs on startup: the kind of user that's
interested in running spectrwm will definitely have at least one
terminal emulator installed.

We can drop D03-make-default-programs-optional.diff and get ever
so slightly closer to upstream.