Switch to 'gbp pq' for patch management
We're already using gbp for building, so it makes perfect sense
to use it for patch management as well; moreover, adopting a
standard tool is preferable to maintaining our own custom
scaffolding, however simple the latter might be.

Thanks to this change we can stop tracking the top-level .pc/
directory, which was never a good idea in the first place.