control: Drop unused Build-Depends
libxtst-dev is not used. It probably was in the past, but
that's no longer the case today.

libxcb-xfixes0-dev was added in 571cf3956bf1 to work around
a packaging issue in libxcb-xinput-dev, which at the time
didn't have an explicit dependency on it despite its .pc file
asking for it through its Requires.private. The issue has
since been addressed, so we no longer need the dependency.