rules: Install xdg-desktop-portal configuration
GTK is as good a choice as any. It matches the fallback value
that will be used if no configuration is present.

This is needed to take care of #1051018, but doesn't solve
the issue on its own: we also need spectrwm to identify itself
via its xsession file. That change has already been merged
upstream and will be part of the upcoming 3.6 release.