control: Drop Depends on libxt6
We don't actually link against it, we just intercept one of
its functions via libswmhack. If the library is not present
on the system, by definition none of the applications that
use it will be able to start, and there will be nothing to
intercept. So a dependency on our side is not useful, and it
just gets in the way of the ongoing 64-bit time_t transition.

Closes: #1068203